refactor: extract ProcessMonitoringCoordinator from AppModel (Phase 4)
Move process monitoring, session attachment reconciliation, synthetic
Claude session management, and all process matching logic into a
dedicated ProcessMonitoringCoordinator (~630 lines extracted). AppModel
drops from 1747 to 1118 lines. The coordinator uses the same
closure-based state access pattern as SessionDiscoveryCoordinator.
